+++ src/julia.h
@@ -261,7 +261,7 @@ typedef union __jl_purity_overrides_t {
         uint8_t ipo_inaccessiblememonly : 1;
         uint8_t ipo_noub                : 1;
     } overrides;
-    uint8_t bits;
+    uint16_t bits;
 } _jl_purity_overrides_t;
 
 // This type describes a single function body
